Rename visdbg to jupedsim_visualizer
