+++ cmake_templates/unix_env.in
@@ -1,2 +1,2 @@
-export PYTHONPATH="@CMAKE_SOURCE_DIR@/python_modules/jupedsim:@CMAKE_SOURCE_DIR@/python_modules/visdbg:@CMAKE_LIBRARY_OUTPUT_DIRECTORY@"
+export PYTHONPATH="@CMAKE_SOURCE_DIR@/python_modules/jupedsim:@CMAKE_SOURCE_DIR@/python_modules/jupedsim_visualizer:@CMAKE_LIBRARY_OUTPUT_DIRECTORY@"
