+++ python_modules/jupedsim_visualizer/bin/jupedsim_visualizer
@@ -2,7 +2,7 @@
 import sys
 
 from PySide6.QtWidgets import QApplication
-from visdbg.main_window import MainWindow
+from jupedsim_visualizer.main_window import MainWindow
 
 if __name__ == "__main__":
     app = QApplication(sys.argv)
