+++ python_modules/jupedsim_visualizer/jupedsim_visualizer/__init__.py
@@ -1,0 +1,7 @@
+import jupedsim_visualizer.config
+import jupedsim_visualizer.geometry
+import jupedsim_visualizer.geometry_widget
+import jupedsim_visualizer.grid
+import jupedsim_visualizer.main_window
+import jupedsim_visualizer.move_controller
+import jupedsim_visualizer.view_geometry_widget
