--- python_modules/visdbg/visdbg/__init__.py
@@ -1,7 +1,0 @@
-import visdbg.config
-import visdbg.geometry
-import visdbg.geometry_widget
-import visdbg.grid
-import visdbg.main_window
-import visdbg.move_controller
-import visdbg.view_geometry_widget
